+++ CHANGELOG.md
@@ -6,7 +6,9 @@ _When adding new entries to the changelog, please include issue/PR numbers where
 
 ## Unreleased
 
+- diff/show: Faster output for some large repositories (varies wildly) [#1038](https://github.com/koordinates/kart/issues/1038)
 - show: Added `--no-sort-keys` option to disable sorting of features by name/PK. Previously added to `diff` only
+- show: Added `--add-feature-count-estimate` option to add a feature count estimate to `json-lines` output. Previously added to `diff` only
 
 ## 0.16.1